requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
autobahn==24.4.2
Automat==24.8.1
boto3==1.34.84
botocore==1.34.84
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
dataclasses-json==0.6.4
decorator==4.4.2
defusedxml==0.8.0rc2
distro==1.9.0
Django==4.2.11
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
django-ses==3.6.0
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
et-xmlfile==1.1.0
exceptiongroup==1.2.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==7.1.0
incremental==24.7.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.42
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langsmith==0.1.105
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
Markdown==3.6
marshmallow==3.21.1
moviepy==1.0.3
msgpack==1.0.8
multidict==6.0.5
mutagen==1.47.0
mypy-extensions==1.0.0
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.17.1
openpyxl==3.1.2
orjson==3.10.0
packaging==23.2
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.2
proglog==0.1.10
psycopg==3.1.18
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
PyJWT==2.8.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytube==15.0.0
pytubefix==6.13.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.1
scikit-learn==1.5.1
scipy==1.14.1
service-identity==24.1.0
setuptools==74.0.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
soundfile==0.12.1
soxr==0.5.0
SQLAlchemy==2.0.29
sqlparse==0.4.4
tenacity==8.2.3
threadpoolctl==3.5.0
tiktoken==0.6.0
tqdm==4.66.2
Twisted==24.7.0
txaio==23.1.1
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
websockets==13.0
whisper==1.1.10
xlrd==2.0.1
yarl==1.9.4
yt-dlp==2024.8.6
zipp==3.18.1
zope.interface==7.0.3
